Revert "Cast ApkAssets to uintptr_t before jlong"
This reverts commit 1d46f4f2dd549bb8711ed12d3c638a9fbef5a18f.

Reason for revert: Did not fix b/159041693

+++ core/jni/android_content_res_ApkAssets.cpp
@@ -354,7 +354,7 @@ static void NativeDestroy(void* ptr) {
 }
 
 static jlong NativeGetFinalizer(JNIEnv* /*env*/, jclass /*clazz*/) {
-  return static_cast<jlong>(reinterpret_cast<uintptr_t>(&NativeDestroy));
+  return reinterpret_cast<jlong>(&NativeDestroy);
 }
 
 static jstring NativeGetAssetPath(JNIEnv* env, jclass /*clazz*/, jlong ptr) {
